@@ -1,18 +1,17 @@
-import Vue from 'vue';
+import { createApp } from 'vue'
 import App from '@/App.vue';
 import {axios, API} from "@/rest_api";
 
 // 导入element ui
-import ElementUI from 'element-ui';
-import 'element-ui/lib/theme-chalk/index.css';
+import ElementPlus from 'element-plus';
+import 'element-plus/lib/theme-chalk/index.css';
 
-Vue.prototype.$axios = axios;
-Vue.prototype.$rest_api = API; // 将API接口包注册到Vue全局
+const app = createApp(App);
 
-Vue.use(ElementUI);
+app.use(ElementPlus);
 
-Vue.config.productionTip = false
+app.config.globalProperties.$axios = axios;
+app.config.globalProperties.$rest_api = API; // 将API接口包注册到Vue全局
 
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+// app.config.productionTip = false;
+app.mount('#app');
